api-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
websockets==12.0
grpcio==1.62.1
grpcio-tools==1.62.1
python-dotenv==1.0.1
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.2
httpx==0.25.2